tratamento de acidentes

+++ Aeda_Proj1/AcidentesViacao.cpp
@@ -2,7 +2,7 @@
 #include "AcidentesViacao.h"
 
 
-AcidenteViacao::AcidenteViacao(string localAcidente, int data, int x, int y, int numFeridosGraves, int numVeiculosEnvolv) : Acidente(localAcidente, data, x, y) {
+AcidenteViacao::AcidenteViacao(string localAcidente, int data, double x, double y, int numFeridosGraves, int numVeiculosEnvolv) : Acidente(localAcidente, data, x, y) {
 	this->numFeridosGraves = numFeridosGraves;
 	this->numVeiculosEnvolv = numVeiculosEnvolv;
 }
