+++ Aeda_Proj1/PostoDeSocorro.cpp
@@ -3,7 +3,7 @@
 
 PostoDeSocorro::PostoDeSocorro(){}
 
-PostoDeSocorro::PostoDeSocorro(int numSocorristas, int numVeiculos, string local, int x, int y) {
+PostoDeSocorro::PostoDeSocorro(int numSocorristas, int numVeiculos, string local, double x, double y) {
 	this->numSocorristas = numSocorristas;
 	this->local = local;
 	this->numSocorristas = numSocorristas;
